requirements.txt
requests
python-dotenv
apscheduler
python-socketio
aiohttp
pytz